+++ src/app/pages/LoginPage.tsx
@@ -1,37 +1,25 @@
 import { FormEvent, useState } from "react";
 import { Typography } from "@mui/material";
-import { useHistory } from "react-router";
 import { useRouteParams } from "../../lib/hooks/useRouteParams";
-import { useLoginMutation } from "../state/client";
-import { loginRedirect, router } from "../router";
+import { router } from "../router";
 import { Link } from "../components/Link";
 import { CenteredContent } from "../components/CenteredContent";
 import { Header } from "../layout/Header";
 import { UserLoginForm } from "../forms/UserLoginForm";
 import { LoginPayload } from "../../api/services/user/types";
-import { useAppDispatch } from "../state/store";
-import { auth } from "../slices/auth";
+import { useLogin } from "../slices/auth";
 
 export default function LoginPage() {
-  const dispatch = useAppDispatch();
   const { destination } = useRouteParams(router.user().login);
   const [loginPayload, setLoginPayload] = useState<LoginPayload>({
     username: "",
     password: "",
   });
-  const { mutateAsync: login, error, isLoading } = useLoginMutation();
-  const history = useHistory();
+  const [login, { error, isLoading }] = useLogin(destination);
 
   async function submit(e: FormEvent) {
     e.preventDefault();
-    let token: string;
-    try {
-      token = await login(loginPayload);
-    } catch {
-      return;
-    }
-    dispatch(auth.actions.update(token));
-    history.push(destination ?? loginRedirect);
+    login(loginPayload);
   }
 
   return (
